The BCD comparison code didn't handle infinities correctly.
In two places, there was code that relied on isinf(3) returning -1 for
negative infinity, but that is not standard behavior (the standards only
require isinf(3) to return a nonzero value). This broke importing negative
infinities when importing a binary state file into a decimal build under
Cygwin; I fixed the code so it doesn't require this behavior.