Another case of the "config" absence blues
There are three cases where the iOS version needs the "config"
directory to exist: when writing state, when presenting the States
window, and when receiving a state. The first of those has always been
handled correctly, the second was fixed a couple of commits ago, and
this commit finally addresses the third.