Apple makes me bump the version number again
I uploaded a build 2.5.3a earlier today, to address the issues with a
missing "config" directory, but I didn't submit it because I found yet
another one. But submitted or not, in Apple land, every upload must
have a higher version number than the last. So here we go: 2.5.3b.