Fixed iOS hostname lookup
The DNS hostname lookup on iOS (for the HTTP Server window) could
access memory after it was freed, which was probably harmless but still
wrong. Also, it was too picky about which interfaces it would consider,
accepting only en0 or en1, instead of accepting anything that starts
with "en".
Also adding some memory debugging settings to the Debug scheme for iOS,
in hopes of catching whatever it is that is causing the mysterious
crashes reported by one user in 2.5.1.