Build script and HISTORY cleanup
Since the main source code package is made obsolete by Free42 being on
GitHub now, I removed the code that builds free42.tgz, and moved the
code that builds the MacPorts and Linux upstream packages to the
build-mac and build-gtk scripts, respectively. All that's left of the
old build-src-doc script is the bit that builds history.html from the
HISTORY file; that is now the build-history script.

Finally, in the aforementioned HISTORY file, I changed all references
to the "iPhone version" to "iOS version", and similarly, "Mac" to
"MacOS", and "MacDashboard" to "MacOS Dashboard". Partly this is just
because I want the terminology to be consistent between the HISTORY
file and the Free42 web site, but there's also the issue of not
confusing operating systems and devices.