Fixed several memory management bugs
There were a couple of bugs where objects were passed from one thread
to another without being retained; one of these was causing crashes
while running programs on an iPhone 8.
The States dialog in the Mac version mis-handled some internal state,
causing private data in the states list to be freed prematurely,
causing crashes if a state was switched to for the second time in a
session.