Moved three settings from core to shell
The settings Singular Matrix Error, Matrix Out Of Range, and Auto-
Repeat are core settings in the sense that they control the behavior of
the core, but from a UI perspective, they belong in the shell; they are
controlled from the Preferences dialog, and should not change when the
user switches states.

Hence, moving those three settings into the shell state. Because I
don't want to create yet another new core state version, I'm leaving
those settings in the core state file as well, but from now on, those
will only be written to, for forward compatibility, and not be read. If
the shell encounters a sufficiently recent shell state file, it will
read the three settings from there; otherwise, they will be initialized
to hard-coded defaults. The settings from the core state will *always*
be ignored.