Fixed program read/write during state load/save
When core_import_programs() and core_export_programs() were called with
NULL for the filename parameter, signalling that they are reading or
writing the state file, the iOS case didn't set raw_buf to NULL, which
could cause the code to assume it's reading from a memory buffer. This
didn't actually seem to cause problems on the real device, but it did
in the simulator, and it was wrong anyway.