iOS crash band-aid; SOLVE bug fix
The iOS version crashes sometimes in [CalcView drawRect:]. The reason
is not yet understood, but forcing the core to exit after just a few
calls to shell_needs_cpu() mitigates the issue. This is a temporary
measure, to be replaced by a proper bug fix once I understand the root
cause of the problem.

SOLVE would sometimes not return the best value in Z. Fixed.