Fixed VARMENU behavior across restart
The store_command() function calls draw_varmenu(), because editing a
program can cause local variables to be deleted, which then needs to be
reflected in the display. However, while loading programs from the
state file, none of the variables have been loaded yet, and this
draw_varmenu() call is unwanted in that case, because when it sees that
there are no variables at all, it will exit the VARMENU altogether.