Made the iOS version single-threaded
Removed all threading from the iOS version, except for the HTTP Server.
(The HTTP Server uses threads for the hostname lookup, the accept loop,
and for each HTTP request.) Working around the lack of event queue
peek-ahead by limiting core runtime to 10 ms per invocation, and
forcing immediate core exit when [CalcView setNeedsDisplay] or
[CalcView setNeedsDisplayInRect:] are called.

If this makes the crashes in [CalcView drawRect:] go away, this could
be a good interim solution until the root cause of those crashes can be
indentified and hopefully fixed. Ultimately, I would prefer to keep
Free42 on iOS multi-threaded because in the absence of event queue
peek-ahead, threads are the only way to maximize throughput without
sacrificing UI responsiveness.