Added click.wav for iOS
Before, the click sound in the iOS version was produced by calling
AudioServicesPlaySystemSound(1105), where 1105 is an undocumented sound
ID that plays the system's keyboard click. Doing it this way had the
advantage of working even on the iPod touch 1st generation, but the
disadvantage that that click was rather quiet, and not responsive to
the volume control. Since recent versions of Free42 don't run on that
device anyway, there was no reason not to use a standard wav file
instead, just like for all the other sounds.