Removing display prioritization
In the single-threaded logic for iOS, I was prioritizing
drawing by making the core return immediately whenever CalcView was
invalidated. Prioritizing drawing like that was a bad idea, since there
is no value in repainting the screen that often, and it has the side
effect of slowing down display-intensive programs by a lot. Fixing this
by removing the display-prioritization code altogether, so the core
will now run for up to 10 ms regardless of what types of events are
pending.