Rename state file on crash
Now renaming the state file from NAME.f42 to NAME.f42.DATE.crash before
reading it, and then back to NAME.f42 afterward. This way, if the app
crashes while reading the state file, it won't crash on that state file
again the next time it is launched, and the state file remains
available for post-mortem debugging.