Added raw file format sanity check for iOS
When importing raw files from links, attachments, or the Files app, now
preventing accidental import of raw images using the same check
implemented for Android in the previous commit.

For iOS 13 users, this will be moot as long as Apple doesn't fix the
Share Sheets that they broke in iOS 13, which they seem to be in no
hurry to do; the bug that prevents apps from being offered as options
in Share Sheets was first reported in June and is still not fixed in
the latest iOS 13.1 beta at the time of writing.
https://forums.developer.apple.com/thread/118932

But it works in 12.4.1 on my old iPhone 5s, which is nice. :-)