Key click and haptic feedback improvements for Android
Added two new, quieter versions of the key click sound, for devices
where the old click sound is too loud. The new versions are like the
old one, except 3 and 6 dB quieter.

For haptic feedback, the slider has the same range as before, 0 to 150,
but now works in steps of 10 instead of 50. (These numbers are the
number of milliseconds the vibrator is run for each key click.)