Key clicks & haptic feedback tweaks
Changed the step size between the key clicks levels from 3 dB to 6 dB,
while maintaining the existing maximum, thus lowering the lowest level
by 24 dB, creating a total dynamic range of 48 dB.

For haptic feedback on Android, changed the linear 0-50-100-150 ms
scale to a logarithmic one of 0 plus sqrt(2)^n, with n from 0 to 15.
This raises the highest level slightly, and adds a lot of resolution at
the lower end of the scale.