Changed Windows file description to "Free42"
It was "HP-42S Calculator Simulator," which makes more sense as a
description, but since the only place most people will ever see this
"description" is in the task bar, in the tooltip that is shown when you
hover over the app icon, showing the app name makes more sense. And
it's what everyone else appears to be doing as well.