Removing the non-selective 'content' filter (Android)
The non-selective intent-filter for scheme="content", i.e. the one that
matched on all MIME types, has been removed. Tired of fielding the
complaints, and there is a reasonable work-around, in that you can save
email attachments to local files, and then open them using a file
manager. It's a bit roundabout, but it will work out better for the
people who don't care about that functionality to begin with. Lesson
learned: file handling in Android stinks, and since those APIs have
been around since Android 1.0, anyone waiting for Google to improve
them is going to be waiting a long time.