Android current state rename fix
When renaming the currently active state in the Android version, the
state name was not updated in the shell settings, with the result that
upon exit, the state would still be saved under the old name, and at
the next launch, would also be loaded under the old name. The net
effect of all this was that instead of ending up with a renamed state,
you'd end up with a copy.