Menus mostly working
Except for Skins. Also, tons of warnings on the console. Clearly, I
don't know what I'm doing, but it is also odd that the warnings sound
like the menus shouldn't be working at all, and yet they do.
And of course, still need to implement the new expose event handling,
and single-instance behavior using GtkApplication.