Added note about GTK deprecation warnings
It looks like most, maybe all, of those warnings are about functions
that were deprecated in GTK 3.22, when new functions to replace them
were introduced. For now, I'm going to leave things the way they are,
since moving to a GTK 3.22 build environment would result in binaries
that won't run on older distros.

My current build environment is Ubuntu 12.04 with GTK 3.4.2, producing
binaries that run on a lot of systems; GTK 3.22 was released in 2016
and thus isn't present on many older systems.