Thin-line fix for GTK version
In scaled mode (GDK_SCALE greater than 1), bitmap smoothing causes
drawing of the display bitmap to bleed outside the display rectangle,
and this can lead to thin dark lines being left behind around the
display, and that looks ugly. A similar problem exists in the Android
version, and I implemented a similar fix here, namely, to always
repaint the entire display when the core updates it, and to update
a slightly larger rectangle than the display itself.

While this might be expected to make things slower, it actually
doesn't, because display repainting now happens *only* in response to
exposure events, so it actually happens less often and programs run,
if anyhting, faster than before. Because of the lack of downsides,
the new logic is permanent, and not a Preferences option.