Made new GTK display logic optional
While the new logic is faster in Ubuntu 18.04, Ubuntu 19.10, and
FreeBSD 12.1, using both GDK_SCALE=1 and GDK_SCALE=2, it turns out
that it is a lot *slower* in Ubuntu 12.04. So, the old logic should
still be an option for people using older systems. I am making the
new logic the default, though.