GTK display repaint fix
Contrary to what I thought earlier, painting with GDK_SCALE > 1 does
not paint outside the destination rectangle. The thin black lines are
caused by partial transparency applied just *inside* the destination
rectangle, and this means that always repainting the entire display is
not necessary in this scenario; it is sufficient to paint the
destination rectangle with the display background color before blitting
the display bitmap. Fixed the "old logic" accordingly.

Also, changed the "old display logic" option to "always repaint entire
display", with inverted sense (so no change to the state file). This
behavior is now the default once more, but note that I'm not removing
the "always repaint entire display" option, since it may still be
necessary if non-integral scaling is in effect, or if some GTK versions
behave differently than the one I'm testing with.