GTK repaint cleanup
Changing repaint logic to only paint in response to draw events, never
unsolicited. The latter doesn't appear to work in Wayland, and I'm not
sure if that's a bug in the Wayland back-end for GTK or something
you're not supposed to do anyway. Since painting only in response to
draw events is more efficient, it makes sense to always do this, in
which case I may as well get rid of the old logic. For now, it's
conditional on the static bool no_unsolicited_repaints, which is
currently hard-coded to be 'true'.