Text smoothing fix for scaled mode in GTK 3
On some platforms, when running with GDK_SCALE greater than 1, there
would be a thin dark line around the display area. This was caused by
bitmap smoothing, where apparently a poor smoothing algorithm would
assume that the area immediately outside the display bitmap was black,
and then blend that black into the pixels at the edge. (Smarter
platforms consider the area outside the bitmap to be transparent, and
bleed partial transparency into the bitmap, rather than blackness.)

The fix consists of adding two pixels around the edge of the display
bitmap on all sides; one pixel is needed to allow for the black pixels
inside the display to bleed one pixel outside, and one additional pixel
is needed to always be the display background color. When updating the
screen, that additional, outermost pixel is masked off; this way, the
text is properly smoothed even at the edges, while no black from
outside the edges can bleed in.  N.B. On the Raspberry Pi running
Raspbian 10, it is necessary to turn on "always repaint entire display"
mode in addition to this new hack; in Ubuntu 19.10 running in
VirtualBox, neither the new hack nor "always repaint entire display"
are needed. The new hack is always on since its cost is essentially
zero; the "always repaint entire display" hack remains optional because
its cost can be significant, at least while running display-intensive
programs.