Android versionCode changes
Bumping the versionCode, because I have to, in order to resubmit the
app to see if excluding below-13 from the age groups gets it past the
"Designed for Families" hurdle. (Google's docs aren't very clear on
this point so it looks like I have to engage in a bit of trial and
error here.)

Also taking the opportunity to streamline the way versionCode is
handled; not using a separate version.code file any more, but getting
the current value from build.gradle instead.