Added some Junk of Historical Interest
A few items that may or may not ever be used again, but that I would
hate to lose regardless. The util directory in the source code repo
seems like an appropriate place for them, since that is where I have
been collecting various other odds and ends as well.