Moved 'util' to a separate repository
It was just cluttering things up in this repository. 'util' is really
an archive; it does not contain anything that is required to build any
of the Free42 executables or source packages. So, it makes sense to
move it to its own repository. And this way, source packages created by
simply grabbing this whole repo will be a more reasonable size, without
the bloat of all those large images under 'icon' and 'screen-shots'.