Swipe-to-switch tweaks
Added a 50 ms delay between a touch event registering and acting on it.
If a pan gesture begins during that delay, the touch is ignored. This
prevents swipes that start on a key from pressing that key, while
introducing a delay that is short enough not to be annoying. N.B. I
also tried 10 ms, and that was too short.

Also, when deciding whether to flip views or not, now extrapolates the
translation by adding a bit of velocity to it, to allow short but quick
swipes to do the job. Seems to work well.

More testing will be needed to see whether these changes feel good over
a longer period of time.