No really, release 2.5.17
My Mac notarization scripts stopped working. Long story, which is still
not resolved. Using Xcode for now, which is a pain in the rear compared
to running a few scripts. Had to add "hardened runtime" to get
Archiving and Notarizing to work in Xcode, but that doesn't seem to
have fixed the scripts. No idea what else might be wrong yet.