Swipe-to-switch for Android, phase 1
Not quite as nice as the iOS implementation, since this is based on
fling events rather than panning. But a nice implementation in Android
looks like a serious PITA; as usual for Android, the standard APIs just
aren't very helpful.

Having said all that, the swipe approach is certainly usable. It just
needs a bit of extra logic to delay DOWN events on the calc view
slightly, so that we can suppress them if a fling is recognized within
the delay interval. That part will be similar to what I had to do in
the iOS version. With that done, and a bit of testing on a couple of
real phones, this should be fit for general release.