Fixing gaps in a couple of skins
Gave the ancient standard skins for mobile the max_sens_rect treatment.

I thought I had already done this for all the skins in my collection,
but these two, being in the source tree and not in the main skin
collection, were apparently missed back then. While testing the latest
Android version by playing a few games of MINEHUNT, I noticed just how
annoying it was that it was missing lots of keystrokes due to me not
hitting the keys accurately enough, and that was just not acceptable.