Removing swipe-to-switch for Android
It is badly broken, apparently at least in part due to the swipe
detector looking at swipe distance in terms of logical pixels instead
of physical screen distance, which makes it hyper-sensitive when high-
resolution skins are in use. I'll have to go back to the drawing board
for this feature, and implement pan detection like in iOS. Android
doesn't appear to support that, so I may have to deal with event
handling myself.

I had also yanked n-key rollover from the emergency release 2.5.17b
(not committed since it was just a rollback to 91234eb), but after
third-party testing, it looks like that was not the cause of the
keyboard problems, so I'm keeping it in this commit, and it will
reappear in the next release.