Removing n-key rollover from Android version
Even though testing indicates that the UI issues in 2.5.17a were caused
by the swipe recognition, not the n-key rollover, I've decided to get
rid of all the Android UI changes, and remove the swipe detection *and*
the n-key rollover. The Android event handling API is too much of a
minefield for my taste, and since the severe issues in 2.5.17a never
showed themselves on the Moto G5 that I use for development and
testing, I am no longer confident of being able to test any Android
event handling changes.

So, swipe-to-switch and n-key-rollover will remain iOS-only. This
commit makes official what is already deployed in the Play store.