Fixed Copy to Free42 behavior for iOS
It had developed a tendency to import files twice, because the check
that was there to recognize that the url parameter to
[Free42AppDelegate application:openURL:options:] wasn't also imported
from $HOME/Documents/Inbox had stopped working, because now all of a
sudden the path in the url is under /private while the $HOME path is
not. Even though they are still pointing to the same directory.

I guess I should rethink that logic. Why am I even scanning the Inbox
and not just looking at the url parameter and nothing else? And if I
must do both, shouldn't I make the urlInInbox check more robust by
using stat() and making sure the parent of the url parameter is the
Inbox directory by looking at its inode number or something?