Finished compensated DOT
Moved the dot product calculation to three separate functions, for the
real-real, real-complex, and complex-complex cases; changed docmd_dot()
to use these three functions. The objective of putting the
implementation of the compensated dot product in these separate
functions is to allow them to be used in other functions as well, e.g.
LU decomposition, matrix-matrix multiplication, etc., but that part
will be done at some later point in time.