Added True Heading support for Android
This completes the Android HEADING update. Tested on Samsubg Galaxy A6.
Note that the new HEADING will not return true heading by default; this
requires LOCAT to be executed first, so the device starts receiving
location updates, since these are required to get the geomagnetic field
data. (On iOS, this data is hidden behind the heading API and does not
require location permissions, so true heading is returned by default on
that platform.)