iOS build tweaks
Changed deployment target from iOS 8.0 to 9.0. The former is no longer
in the list of options in Xcode; it was still working but it's not
really necessary to stick with that setting since all the devices that
can run 8.0 can run 9.0 as well. This change does lead to a bunch of
deprecation warnings about UIActionSheet, UIAlertView, and
sendSynchronoutRequest:returningResponse:error:. I guess I'll have to
deal with those at some point, and I should probably deal with the
warnings about launch images as well at that time.

Changed libz.dylib to libz.tbd.