Update G & RAD annunciators in STOFLAG
For obvious reasons...

Also, fixed handling of rad and grad flags. The case where both are set
was handled inconsistently, sometimes treated as rad and sometimes as
grad. This wasn't an issue before, but due to the existence of STOFLAG,
it is now. This case is consistently handled as RAD now.