Removed extension disabling logic
It was messy, somewhat broken, and I've never used it and can't think
of any reason why I would want to at this point, having done fine
without it for 16 years. All that remains is the logic to hide FPTEST
from the FCN catalog for non-debug builds, and ACCEL, LOCAT, and
HEADING from non-mobile builds.