Improved Export filename generation for Windows
This is still pretty lame, since it uses the ANSI APIs, and it really
needs to use the Unicode APIs to handle the entire HP-42S character set
properly. However, even in its current form, this is better than
before, because it was just using the UTF-8 returned by
core_list_programs() and treating it as ANSI.

How hard using Unicode will be is still to be determined. I tried using
SendMessageW() instead of SendMessage() to populate the program
selection dialog, but the strings still ended up being converted to
ANSI, as if the dialog itself were in ANSI mode. Can I fix that by
creating the dialog using DialogBoxW() instead of DialogBox()? If so,
that's the way to go. If not... if the only way to make this work right
is to build the entire UI with #define UNICODE, that would be a major
chore.