HTTP Server networking fixes
The search for a network interface to bind to, and the initial name
lookup, are now performed when the HTTP Server view is presented, not
at app startup. This has two advantages: first, it is now able to
handle network configuration changes while the app is running, and
second, the permissions dialog posted by the OS to warn the user that
the app is trying to access the network (and causing them to worry what
kind of nefarious activity the app is getting up to) is now presented
within the context of starting the HTTP Server, rather than out of the
blue when the app starts. The possibility of a slow hostname lookup is
now handled better as well, by simply updating the displayed hostname
whenever it becomes available.