Fixing skin & keymap constness for GTK and Windows
The compiled-in keymap and skins were not declared properly to make
them const. This is because way back when I wrote that code, I didn't
know how to do that so that it works across compilation units,
specifically, that 'const' implies 'static' so you have to say 'extern
const' to force the object to be exported. Whew.  Fixed at last.