Fixed name clash in Windows binary debug build
The Intel build script includes a few binary functions that older
versions of VC++ didn't have, but VS 2019 does. These functions are
thus colliding with the VS-provided versions. The decimal build is not
affected because it doesn't use those functions, so the conflict
doesn't materialize, and the binary release build is also not affected,
possibly because the functions are mapped onto FPU instructions.
Fixed the conflict by removing the offending functions from the library
build.