Big Stack
Incomplete and untested. It does compile, but that's all I'll say for
it at this point. Still needs FUNC to be updated to support the big
stack, and the LNSTK and L4STK, for local (temporary) big stack mode
setting, are not yet implemented. Also, the behavior of the [ENTER] and
[<-] keys needs to be different in big stack mode.