No bug-compatible stack quirks in big stack mode
The weird stack lift behavior of EDIT, EDITN, and MATA/MATB/MATX is
faithfully reproduced in Free42 for the sake of compatibility, but in
big stack mode, we don't have backward compatibility anyway, so we may
as well make those functions act in a sensible manner.

Similar considerations apply to the bizarre way in which SUM, MEAN, and
SDEV return their results.