GTK build fixes
Removed the bid128_hypot() patch from intel-lib-unknown-32bit.patch and
intel-lib-unknown-64bit.patch. It shouldn't have been there to begin
with, since these patches are applied *in addition* to
intel-lib-linux.patch, which already contains the bid128_hypot() patch.

Also, added i386 to the list of architectures that causes
intel-lib-unknown-32bit.patch to be applied, in an attempt to get the
build to work on FreeBSD on i386.