VARMENU and R↓N / R↑N fixes
When a function menu was active on top of VARMENU, functions like RCL
would show the VARMENU instead of the variables catalog. That was
wrong, I'm not sure why I coded that, probably a misunderstanding of
how the real 42S works. On the real calculator, the only effect of
VARMENU on functions like RCL is that it is shown instead of the
variables catalog *if VARMENU is on top*. Fixed.

R↓N / R↑N would accept 1 when DEPTH = 0, instead of showing Size Error.
Fixed.