Local INDEX
INDEX now preserves the indexed matrix and IJ, if performed on a matrix
that is local and closer to the top of the stack than the currently
indexed matrix.
Note that the same behavior should also be implemented for EDIT and
EDITN, but those code changes aren't in this commit.