Local EDITN
Same deal as the INDEX change in the previous commit.

I'm not making the same kind of change in EDIT... Not because I don't
want to, but because there is no obvious moment when the IJ restoration
would have to take place. Since the matrix under edit is in X and not
in a local variable, it doesn't get deleted on RTN.