Mystery catalog code cleanup
This is a bit of logic that deals with the situation when ENTER is
pressed, during incomplete alpha command entry, when CATALOG has been
opened on top of the ALPHA menu. The intent of the original logic had
been lost, and after many edits, it didn't make sense any more. I
checked with the real 42S and the point is this: when CATALOG has been
opened on top of ALPHA, you can get back to ALPHA by pressing ENTER.
This is good because pressing EXIT would cancel command entry
altogether.

The list of exceptions was to keep track of the cases when you could NOT
get back from the CATALOG to ALPHA by pressing ENTER, but that could
have been handled much more simply, by reversing the sense of the test
and only coding the catalog sections where you *are* allowed to use
ENTER to get back to ALPHA, and that list is actually quite short: only
the top level.