Allowing L4STK even when "Allow Big Stack' is turned off
Also, DEPTH, DROP, DROPN, DUP, DUPN, R↓N, R↑N, PICK, and UNPICK.

When I first implemented that mode, the idea was to make everything
NSTK-related completely disappear. I later relaxed that to only disable
the big stack functions, not hide them. But the logical consequence of
not hiding those functions is that you might as well enable them, too;
as long as NSTK and LNSTK are disabled, and as long as disabling "Allow
Big Stack" forces 4STK mode, that will prevent people from shooting
themselves in the foot by accidentally running legacy programs in NSTK
mode. And that was the whole point of the exercise anyway.