Fixed ROTXY in the binary version
It used floor() as if it rounded to zero. Which, by coincidence, it did,
in the decimal version, because of the floor(Phloat) bug. But
apparently, the binary version has been getting this wrong since 2004. I
assume nobody ever noticed because applying a negative non-integral
shift count to ROTXY isn't going to be a very common use case...