Logic to show About on launch when upgrading (Mac)
I'll use this for the one-time ad for Plus42.
The About box will have to be stretched a bit, to make room for an
image, showing the Plus42 name and icon, a screen shot highlighting the
big display, and the image should be linked to the Plus42 web page
(desktop versions) or the appropriate app stores (mobile versions).