More efficient GIF loading
Now only switching from monochrome to indexed when encountering a color
that isn't black or white; now only switching from indexed to truecolor
when using more than 256 colors globally. For most skins, this change
doesn't make a difference, but for skins with local colormaps, the
difference can be substantial. For example, mimax3 has 2.5 million
pixels, so whether that is stored as 8-bit indexed or 24-bit truecolor
makes a difference of 5 megabytes.