iOS About box version number fix
Because of the way App Store Connect now interferes with
CFBundleVersion, it has become useless for populating the version in the
about box, and it's irrelevant everywhere else. On the device, the
version number that is shown in the App Storage info is the one from
CFBundleShortVersionString, and in the App Store, it's the version that
was specified when creating the release.
This fix is to populate the version in the About box from
CFBundleShortVersionString, which is still under our full control.