Quick and dirty display speedup for Windows
Don't paint the display in shell_blitter(), only invalidate;
In shell_wants_cpu(), return 'true' no more than once every 10 ms.
This helps a lot, to the point of actually being faster than the
pre-GDI+, pre-fractional scaling logic (3.0.10 and before).
It would be nice to optimize the WM_PAINT handler as well, so if
the dirty rectangle doesn't extend beyond the display rectangle,
it will repaint only the display, and not bother with the skin image
etc. Leaving that for the next commit.