Removed incorrect claim about optimizer bug
In truth, the code in question was wrong: it was written with the
assumption you could shift a 64-bit integer by 64 bits, while in fact,
the C and C++ standards specify that the behavior of the shift operators
is only defined when the shift count is between 0 and word size - 1.
Oops!