Better NN->S
Making it behave like N->S is a bit silly, now that I think about it.
The only reason to use NN->S instead of N->S is because of its behavior
in ALL mode, and so it makes way more sense to have NN->S behave like
it's in ALL mode at all times. And now it does.