Binary: max digits for SHOW increased to 17
This was necessary because 16 digits was actually never enough to begin
with. For example, 5 1/X and 5 SQRT X^2 1/X both return a result that
looks like 0.2 in SHOW, even though they are different, as can be easily
verified with X=Y? or by subtracting them. It takes 17 decimal digits to
ensure that different binary numbers will always be rendered as
different decimal numbers.