Added beep.wav in preparation for async tone playing
This is simply a concatenation of tone8.wav, tone5.wav, tone9.wav, and
tone8.wav. Asynchronous TONE playing is going to be a pain anyway, but
dealing with BEEP being a sequence of four TONEs makes the whole thing
at least twice as annoying. Combining the sequence into one file will
help a lot, even if it adds another 16 kilobytes of raw audio.