Made charset match Plus42
For the sake of copy & paste compatibility between Free42 and Plus42, I
changed eight codes, as follows:

   30: ^
  130: Ω
  131: ᴛ
  132: ʟ
  133: ᴀ
  134: ▹
  136: «
  137: »

None of the original characters with those codes are in the original
ALPHA menu, so you wouldn't find them in program listings anyway. What's
more, all of them have exact duplicates; 30 is identical to 4, and the
others are identical to the corresponding codes in the lower half.
Should these changes cause problems despite all that, I can always add a
toggle to Preferences to restore the original charset.