Added #pragma STDC FP_CONTRACT OFF
Apparently, Apple's ARM compilers default to using FMA for expressions
like a * b - c * d, which causes undesirable results in Free42 and
Plus42. Specifically, (pi-pi*i)*(pi-pi*i) was producing a result with
nonzero real part in the binary builds.
Note that the x86_64 builds, iOS and MacOS, were not affected, and
neither were the Android, Linux and Windows builds.
I'm adding a #pragma to turn off this optimization, rather than the
-ffp-contract=off command-line option, so I won't have to deal with
making this change in all projects and makefiles, and having to keep
track of which compilers need it and which ones don't.