Make ^ type ↑ again
I made the Free42 character set match the one in Plus42 three weeks ago,
and that involved adding a few characters, but it also involved making
the ^ key on the PC keyboard type the ^ character instead of ↑. That
wasn't a good move, because the HP-42S doesn't *have* the ^ character,
so this is a weird change in behavior, even if the most annoying side
effects are fixed by making ^ and ↑ be treated as equal in
find_builtin(). It's well and good to have ^ for text compatibility with
Plus42, but it shouldn't just appear when a user is doing normal HP-42S
things on what they view primarily as an HP-42S simulator.