Fixed completion mechanism for DET and INVRT
Up to 3.3.2, the completion mechanism was implemented inconsistently. In
3.3.3, I fixed the inconsistencies by removing unneeded functionality.
Now I'm restoring that functionality, and fixing the inconsistencies by
making sure the completion routines are *always* called, including on
*all* error paths.
Strictly speaking, this is unnecessary, but I prefer it this way,
because it makes the completion mechanism for DET and INVRT consistent
with the one for linalg_div(). That is, callers should be able to rely
on completion routines always being called, so the completion routines
can be used to perform cleanup operations. The fact that DET and INVRT
don't require such cleanup doesn't mean providing that *capability* is a
bad idea. If nothing else, I think this makes the code more readable and
maintainable.