Android: perform cleanup in onDestroy()
That call was already there, actually, but it was commented out. It
looks like I assumed it was redundant, as in, onDestroy() only gets
called before the process exits, but that is of course incorrect. When
you exit the app using Shift-EXIT, for instance, onDestroy() gets
called, but the process is not killed.